Aws/dht11/requirements.txt
boto3
Adafruit_DHT
pyyaml
